refactor: split the text engine into measure, wrap, fit and draw
The split the plan describes, deferred twice on purpose: in 4a because a
280-line file did not need it, and in 4d so that the behaviour diff stayed
about behaviour. Wrapping has arrived and the file is 604 lines, so it is
due.

  measure.ts  the font cache, line measurement, font scaling
  wrap.ts     breaking text to a width, and paragraphs to lines
  fit.ts      the overflow decisions: hidden, fit, ellipsis, truncation
  draw.ts     the TextBlock, its placement and the clip
  index.ts    the public surface, and the two block builders

No behaviour change: `src/text` resolves where `src/text.ts` did, so no
caller moved, and the 66 fixtures and 80 unit tests pass unchanged.
